implement review comments
