+++ tests/test_plugins_os_unix_log.py
@@ -1,9 +1,6 @@
 from datetime import datetime, timezone
 
-from dissect.util.ts import from_unix
-
 from dissect.target.plugins.os.unix.log.atop import AtopPlugin
-from dissect.target.plugins.os.unix.log.audit import AuditPlugin
 from dissect.target.plugins.os.unix.log.btmp import BtmpPlugin
 from dissect.target.plugins.os.unix.log.lastlog import LastLogPlugin
 from dissect.target.plugins.os.unix.log.wtmp import WtmpPlugin
@@ -100,28 +97,3 @@ def test_atop_plugin(target_unix, fs_unix):
     assert bool(results[0].wasinactive) is False
     assert results[0].container == ""
     assert str(results[0].filepath) == "atop_20221111"
-
-
-def test_audit_plugin(target_unix, fs_unix):
-    data_file = absolute_path("data/unix/logs/audit.log")
-    fs_unix.map_file("var/log/audit/audit.log", data_file)
-
-    target_unix.add_plugin(AuditPlugin)
-    results = list(target_unix.audit())
-
-    assert len(results) == 4
-
-    result = results[0]
-    assert result.audit_type == "SYSCALL"
-    assert result.ts == from_unix(1364481363.243)
-    assert result.audit_id == 24287
-    assert (
-        result.message
-        == 'arch=c000003e syscall=2 success=no exit=-13 a0=7fffd19c5592 a1=0 a2=7fffd19c4b50 a3=a items=1 ppid=2686 pid=3538 auid=1000 uid=1000 gid=1000 euid=1000 suid=1000 fsuid=1000 egid=1000 sgid=1000 fsgid=1000 tty=pts0 ses=1 comm="cat" exe="/bin/cat" subj=unconfined_u:unconfined_r:unconfined_t:s0-s0:c0.c1023 key="sshd_config"' # noqa E501
-    )
-
-    result = results[1]
-    assert result.audit_type == "CWD"
-    assert result.ts == from_unix(1364481363.243)
-    assert result.audit_id == 24287
-    assert result.message == 'cwd="/home/shadowman"'
